# openapi-v2.yaml
openapi: 3.1.0
info:
  title: eLearning Template Engine API
  description: |
    Backend API for the composable component-based eLearning authoring platform.
    Supports 89+ component types, per-interaction audio, page completion tracking,
    weighted scoring, full theming (layout + color system), and SCORM 1.2/2004 export.
  version: 2.0.0
  contact:
    name: eLearning Platform Team
servers:
  - url: /api/v1
    description: API v1 namespace
...
paths:
...
  /export:
    post:
      tags: [Export]
      operationId: exportCourseFromJson
      summary: Export SCORM package from raw JSON
      description: Accepts course JSON (legacy or new format), validates, and returns SCORM ZIP
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExportRequest"
      responses:
        "200":
          description: SCORM package ZIP
          content:
            application/zip:
              schema:
                type: string
                format: binary
          headers:
            Content-Disposition:
              schema:
                type: string
                example: attachment; filename="course-slug_scorm.zip"
            X-Export-Warnings:
              schema:
                type: string
              description: Semicolon-separated export warnings
        "422":
          $ref: "#/components/responses/ValidationError"
  /export/validate:
    post:
      tags: [Export]
      operationId: validateForExport
      summary: Validate course data for export
      description: Validates component types, audio refs, scoring config without generating package
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExportRequest"
      responses:
        "200":
          description: Validation result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationResult"
...
  /export/scorm/{courseId}:
    parameters:
      - $ref: "#/components/parameters/CourseId"
    post:
      tags: [Export]
      operationId: exportPersistedCourse
      summary: Export SCORM package from a DB-persisted course
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScormExportOptions"
      responses:
        "200":
          description: SCORM package ZIP
          content:
            application/zip:
              schema:
                type: string
                format: binary
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"
  /export/status/{exportId}:
    parameters:
      - name: exportId
        in: path
        required: true
        schema:
          type: string
    get:
      tags: [Export]
      operationId: getExportStatus
      summary: Get export job status
      responses:
        "200":
          description: Export status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExportStatusResponse"
        "404":
          $ref: "#/components/responses/NotFound"

  # ── Media ───────────────────────────────────
...
components:
...
  parameters:
    CourseId:
      name: courseId
      in: path
      required: true
      schema:
        type: string
        pattern: "^[a-zA-Z0-9_-]+$"
      description: Course identifier
...
  responses:
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    ValidationError:
      description: Validation failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ValidationErrorResponse"

  # ── Schemas ─────────────────────────────────
  schemas:
...
    ErrorResponse:
      type: object
      required: [detail]
      properties:
        detail:
          type: string
    ValidationErrorResponse:
      type: object
      required: [detail, errors]
      properties:
        detail:
          type: string
        errors:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
    ValidationResult:
      type: object
      required: [valid]
      properties:
        valid:
          type: boolean
        errors:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
        warnings:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
...
    ExportRequest:
      type: object
      required: [course]
      properties:
        course:
          type: string
          description: JSON string of course data
    ScormExportOptions:
      type: object
      properties:
        format:
          type: string
          enum: [scorm_1_2, scorm_2004]
          default: scorm_1_2
        includeMedia:
          type: boolean
          default: true
...
    ExportStatusResponse:
      type: object
      properties:
        exportId:
          type: string
        status:
          type: string
          enum: [pending, processing, completed, failed]
        progress:
          type: number
        downloadUrl:
          type: string
          nullable: true
        error:
          type: string
          nullable: true

    # ── Media ──────────────────────────
...